+++ ikii-back-commons/src/main/java/mx/ikii/commons/payload/request/business/BusinessRequest.java
@@ -11,7 +11,7 @@ public class BusinessRequest {
 	
 	private String image;
 	
-	private String category;
+	private String categoryId;
 	
 	private String description;