+++ ikii-back-commons/src/main/java/mx/ikii/commons/payload/response/business/BusinessResponse.java
@@ -11,7 +11,7 @@ public class BusinessResponse {
 	
 	private String image;
 	
-	private String category;
+	private String categoryId;
 	
 	private String description;