+++ ikii-back-commons/src/main/java/mx/ikii/commons/persistence/collection/Customer.java
@@ -17,25 +17,32 @@ import lombok.Data;
  * This class is used as the entity for the Customer resource and represents
  * the Customer collection in the database
  * 
- * @author Francisco Javier Martínez Arazo
+ * @author Francisco Javier Martï¿½nez Arazo
  *
  */
 @Data
 @Document(collection = "Customer")
 @JsonInclude(Include.NON_NULL)
 public class Customer implements Serializable{
-	
 	private static final long serialVersionUID = -5507907273019766759L;
 	
 	@Id
 	private String id;
+	
 	private String name;
+	
 	private String lastName;
+	
 	private String secondLastName;
+	
 	private String email;
+	
 	private String phoneNumber;
+	
 	private String password;
+	
 	private String birthday;
+	
 	private Boolean isEnabled;
 	
 	@DBRef