+++ ikii-back-microservices/ikii-back-business/src/main/java/mx/ikii/business/controller/IBusinessCategoryController.java
@@ -23,7 +23,7 @@ public interface IBusinessCategoryController {
 	@GetMapping("{id}")
 	ResponseEntity<BusinessCategoryResponse> getById(String id);
 
-	@GetMapping("/business/{name}")
+	@GetMapping("/name/{name}")
 	ResponseEntity<BusinessCategoryResponse> getByBusinesName(String name);
 
 	@PostMapping