+++ ikii-back-microservices/ikii-back-business/versionHistory.md
@@ -8,5 +8,11 @@
 
 * renaming of misnamed attributes
 
+* fix path endpoint getByBusinesName
+
+* fix update business
+
+* fix update business category
+
 ## 1.0.0
 * Base ikii proyect